em-dt-causal-4nodes-backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
cffi==1.15.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.103.2
gevent==22.10.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib-metadata==6.7.0
numpy==1.21.6
pandas==1.1.5
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
websocket==0.2.1
wsproto==1.2.0
zipp==3.15.0
zope.event==5.0
zope.interface==6.2
